Move IOHandlers tests to separate dir.
